supress windows warning about ulonglong->ulong in TrackTimingMonitor
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@53823 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: cbef1738df19bb5298a9a3cca2b481e9a07913f8

+++ Rec/Tr/TrackMonitors/src/TrackTimingMonitor.cpp
@@ -203,8 +203,8 @@ StatusCode TrackTimingMonitor::execute()
     const LHCb::ODIN* odin = get<LHCb::ODIN> ( LHCb::ODINLocation::Default );  
     theTuple->column("RunNb", odin->runNumber());
 
-    unsigned long hi_num =  (odin->eventNumber() >> 32);
-    unsigned long low_num = (odin->eventNumber() & 0xffffffff);
+    unsigned long hi_num =  (unsigned long) (odin->eventNumber() >> 32);
+    unsigned long low_num = (unsigned long) (odin->eventNumber() & 0xffffffff);
  
     theTuple->column("EventNb_hi", hi_num);
     theTuple->column("EventNb_low", low_num);
